Added file 102-print_comb5.c and deleted temporary file

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -10,7 +10,7 @@ int main(void)
 	int i;
 
 	for (i = 48; i < 58; i++)
-		putchar((int)i);
+		putchar(i);
 
 	putchar('\n');
